% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mean.R
\name{fct_mean}
\alias{fct_mean}
\title{FUNCTION TITLE}
\usage{
fct_mean(.df, .colnum, .rounding = 0)
}
\arguments{
\item{.df}{A table with at least a numeric column.}

\item{.colnum}{A column name of the input table.}

\item{.rounding}{Rounding number of digits, defaults to 0.}
}
\value{
the mean of the input column values.
}
\description{
Take a numeric column from a dataframe and return it's mean rounded
to the specified number of digits.
}
\examples{
fct_mean(.df = iris, .colnum = Sepal.Width, .rounding = 1)

}
