% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_readzip.R
\name{fct_readzip}
\alias{fct_readzip}
\title{Read OpenForis Arena OLAP zipfile data}
\usage{
fct_readzip(.path)
}
\arguments{
\item{.path}{A path to the data ZIP file.}
}
\value{
a list of entity level data frames and survey descriptors
}
\description{
Processing chaina of OpenForis Arena surveys generate a ZIP file
containing field entity data at the minimum area unit (OLAP tables) and
survey descriptors. fct_readzip() is tailored to load this data.
}
\examples{
zipfile <- system.file("extdata/OLAP_shiny_demo.zip", package = "arenalytics")
zipdata <- fct_readzip(.path = zipfile)
summary(zipdata$OLAP_tree$tree_biomass_ag)

}
