% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run_arenalytics.R
\name{shiny_run_arenalytics}
\alias{shiny_run_arenalytics}
\title{Analytical Dashboard for OpenForis Arena}
\usage{
shiny_run_arenalytics(...)
}
\arguments{
\item{...}{arguments to pass to shinyApp}
}
\description{
function to launch the ShinyApp Arena Analytics. This app helps users to explore
their data from OpenForis Arena, gain insights and develop report-ready figures and
statistics.
}
\examples{
if (interactive()) {

shiny_run_arenalytics()

}

}
