% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_checkzip.R
\name{fct_checkzip}
\alias{fct_checkzip}
\title{Check OpenForis Arena OLAP ZIP file integrity}
\usage{
fct_checkzip(.path)
}
\arguments{
\item{.path}{a path to the zip file from}
}
\value{
A list with TRUE/FALSE.
}
\description{
fct_checkzip() checks if the ZIP files uploaded in the master ShinyApp has
the core files required to run the analysis.
}
\examples{
zipfile <- system.file("extdata/OLAP_shiny_demo.zip", package = "arenalytics")
zip_check <- fct_checkzip(.path = zipfile)
zip_check$all_ok

}
