% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run_arenalytics.R
\name{shiny_run_arenalytics}
\alias{shiny_run_arenalytics}
\title{TITLE}
\usage{
shiny_run_arenalytics(...)
}
\arguments{
\item{...}{arguments to pass to shinyApp}
}
\description{
Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec vehicula
imperdiet finibus. Sed urna sem, molestie at sodales non, viverra vitae
mauris. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque
convallis tristique mauris, nec volutpat ligula dapibus eget.
}
\examples{
if (interactive()) {

shiny_run_APPNAME()

}

}
